♻️refactor: competition workbook 연관관계 추가
